LayoutTests/fast/html/transient-activation-expected.txt

Tests that transient activation is generated by correct events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
eventSender.mouseMoveTo(50, 50)
PASS EVENT(mousemove)
PASS navigator.userActivation.isActive is false
eventSender.mouseDown()
PASS EVENT(mousedown)
PASS navigator.userActivation.isActive is true
eventSender.mouseUp()
PASS EVENT(mouseup)
PASS navigator.userActivation.isActive is false
UIHelper.rawKeyDown('escape')
PASS EVENT(keydown)
PASS navigator.userActivation.isActive is false
UIHelper.rawKeyUp('esc')
PASS EVENT(keyup)
PASS navigator.userActivation.isActive is false
UIHelper.rawKeyDown('a')
PASS EVENT(keydown)
PASS navigator.userActivation.isActive is true
UIHelper.rawKeyUp('a')
PASS EVENT(keyup)
PASS navigator.userActivation.isActive is false
